pyproject.toml
[tool.poetry]
name = "coursework 6"
version = "0.1.0"
description = "Backend for a site with ads"
authors = ["Aleksey Lesnikov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
python-dotenv = "^0.21.0"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.5"
djoser = "^2.1.0"
pillow = "^9.3.0"
django-cors-headers = "^3.13.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.0.0"}
drf-spectacular = "^0.24.2"
drf-nested-routers = "^0.93.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
